Decrease perceived flash of unstyled content time. Fix bug preventing re-selection of tokens containing '+'.
This update separates the initialization of the Select2 field from the loading of the field with data. The field is initialized without data, and once the AJAX calls to each data source are complete the field is loaded. A RegExp escape function from stackoverflow.com/a/6969486/151312 was added to prevent accidental interpretation of RegExp characters when attempting to identify tokens to pre-select.