Implement caching of Select2 data.
To reduce AJAX calls to the server, the results of each category's AJAX calls are stored in the browser's local storage, where they are drawn from on subsequent page views. A cache expiry value is set alongside the storage data, which is configurable in portal-select2.js; it is currently set to 5 minutes.