Replace individual category search GET parameters with unified search parameter.
Previously, searches within a cateogry could be specified using a GET parameter corresponding to that category (e.g., ?d=Melanoma to search for disease=Melanoma). These have been replaced with the 's' parameter for unified searches, with corresponding category tags added to the end of queries (e.g., ?s=Melanoma[disease]). A bug causing 500 errors when an incorrect category was specified is fixed.