Restructure DB to abstract Feature and Attribute definitions. Rewire assertion add/approve pages to work with new DB. Rest of site requires rewiring.
In this change, the DB has been substantially modified to remove the feature/alteration/alt_type columns within Assertion and to instead use a more abstract Feature/Attribute system. Instead of "shoehorning" different types of data within the alteration/alt_type columns depending on the feature, now various FeatureDefinitions and AttributeDefinitions are defined at instantiation of the database. Features may have as many or as few Attributes as necessary to describe the feature type.
Most of the Almanac is broken in this commit due to the DB restructuring. The assertion add/approve pages have been updated to work with the new model in order to allow testing. The following commits will be focused on bringing the remainder of the site up-to-date with the new DB model.