Fix bug preventing carry-through of therapy name from add page to approve page. Fix bug displaying search results when no predictive implication information.
Therapy names were not being carried-forward due to a code typo; Assertion.therapy_type was being set to the submitted therapy name instead of Assertion.therapy_name. In search results, no table cell was being created in the predictive implication column when no predictive implication was set. Now an empty cell is created in the column.