Suggest Attributes and Attribute values when a user begins a Feature search.
When a user is entering a Feature name into the Select2 search field, all corresponding Attributes linked to all existing Attribute:Value pairs is loaded into the suggestions. Additionally removes the automatic prepending of Feature names when a user searches for an Attribute name; users may now search specifically for Attribute:Value pairs independent of Feature, and the search will be performed cross-Feature.