Fix bug preventing correct ordering of search results by predictive implication.
Previously, a separate "Predictive Implication Sort" column was created as part of the search results Data Table, which contained integers corresponding to the Predictive Implication of that row (e.g., "FDA-Approved" = 5). This column was then hidden from view but used to sort the DataTable. The Predictive Implication integers were stored in the database, and automatically inserted on creation by mapping the Predictive Implication string in the source spreadsheet. The new DB model did not include this Predictive Implication column, causing sorting to no longer function. Instead of adding a separate DB column, the "orthogonal data" feature of DataTables was leveraged to allow sorting a column by values separate from each cell's display value. Now, a "data-order" attribute is added to the TD for each Predictive Implication cell, which contains the integer value corresponding to that implication level, and the DataTable is ordered by this column. The logic in db_insert.py for adding information to the Predictive Implication Sort column has been removed.

Additionally, the DataTable column holding the link to each row's assertion has been made non-sortable.